Move more functions to image_functions
