+++ image_functions.py
@@ -2,6 +2,16 @@ import scipy
 import scipy.misc
 from scipy.cluster import vq
 import color_spaces as cs
+import numpy as np
+import Image
+
+def open(target_filename):
+    "Just a wrapper for Image.open from PIL"
+    try:
+        return Image.open(target_filename)
+    except IOError:
+        logger.warning("Cannot open %s as an image.", target_filename)
+        return
 
 def split_regions(img, split_dim):
     """Split an image into subregions.
@@ -43,3 +53,72 @@ def dominant_color(img, clusters=5, size=50):
     counts, bins = scipy.histogram(vecs, len(colors))
     dominant_color = colors[counts.argmax()]
     return map(int, dominant_color) # Avoid returning np.uint8 type.
+
+def plot_histograms(hist, title=''):
+    "Plot an RGB histogram given as a dictionary with channel keys."
+    import matplotlib.pyplot as plt
+    fig, (red, green, blue) = plt.subplots(3, sharex=True, sharey=True)
+    domain = range(0, 256)
+    red.fill_between(domain, hist['red'],
+                     facecolor='red')
+    green.fill_between(domain, 0, hist['green'],
+                       facecolor='green')
+    blue.fill_between(domain, 0, hist['blue'],
+                      facecolor='blue')
+    red.set_xlim(0,256)
+    red.set_ylim(ymin=0)
+    red.set_title(title)
+    fig.show()
+
+def img_histogram(img, mask=None):
+    keys = 'red', 'green', 'blue'
+    channels = dict(zip(keys, img.split()))
+    hist= {}
+    for ch in keys:
+        if mask:
+            h = channels[ch].histogram(mask.convert("1"))
+        else:
+            h = channels[ch].histogram()
+        normalized_h = [256./sum(h)*v for v in h]
+        hist[ch] = normalized_h
+    return hist
+
+def compute_palette(hist):
+    """A palette maps a channel into the space of available colors, gleaned
+    from a histogram of those colors."""
+    # Integrate a histogram and round down.
+    palette = {}
+    for ch in ['red', 'green', 'blue']:
+        integrals = np.cumsum(hist[ch])
+        blocky_integrals = np.floor(integrals + 0.01).astype(int)
+        bars = np.ediff1d(blocky_integrals,to_begin=blocky_integrals[0])
+        p = [[color]*freq for color, freq in enumerate(bars.tolist())]
+        p = [c for sublist in p for c in sublist]
+        assert len(p) == 256, "Palette should have 256 entries."
+        palette[ch] = p
+    return palette
+
+def adjust_levels(target_img, from_palette, to_palette):
+    """Transform the colors of an image to match the color palette of
+    another image."""
+    keys = 'red', 'green', 'blue'
+    channels = dict(zip(keys, target_img.split()))
+    f, g = from_palette, to_palette # compact notation
+    func = {} # function to transform color at each pixel
+    for ch in keys:
+        def j(x):
+           while True:
+               try:
+                   inv_f = f[ch].index(x)
+                   break
+               except ValueError:
+                   if x < 255:
+                       x += 1
+                       continue 
+                   else:
+                       inv_f = 255
+                       break
+           return to_palette[ch][inv_f]
+        func[ch] = j 
+    adjusted_channels = [Image.eval(channels[ch], func[ch]) for ch in keys]
+    return Image.merge('RGB', adjusted_channels)
